[BEAM-1722] Move PubsubIO into the google-cloud-platform module
